Split graphics into another module
